mypy clean, and a real bug the last flake was hiding
mypy: 28 -> 0 across 334 files.

The one that matters most: OrderNumberGenerator declared next_number as
sync while the only implementation is async, and OrderService never
awaited it. Every order number was therefore a coroutine object rather
than a string. Making the port honest surfaced it; the fakes moved too.

Also real: the catalog publish/unpublish routes passed publish=bool to a
service that takes an explicit state, so all three would have raised on
call; and the Mini App auth dependency I added earlier declared User
while from_mini_app returns an optional UserProfile.

Then the intermittent one. test_generated_secrets... failed in the suite
and passed alone, which looked like flakiness. It was not:
generate_secret draws from an alphabet where a 48-character string spells
'todo' about once in 23,000 draws, and weakness_of rejects any value
containing a development marker. So roughly 1% of 200-secret runs - and,
far worse, the occasional real deployment - would get a generated key
that this module's own guardrail refuses at boot. The generator now
re-draws until its own validator accepts, with a bounded loop, and a test
pins the agreement over 5,000 draws.

+++ src/geekvpn/infrastructure/security/throttling.py
@@ -174,7 +174,7 @@ def lockout_seconds(consecutive_failures: int) -> int:
     if consecutive_failures < LOCKOUT_THRESHOLD:
         return 0
     excess = consecutive_failures - LOCKOUT_THRESHOLD
-    return min(LOCKOUT_BASE_SECONDS * (2**excess), LOCKOUT_MAX_SECONDS)
+    return int(min(LOCKOUT_BASE_SECONDS * (2**excess), LOCKOUT_MAX_SECONDS))
 
 
 def requires_captcha(consecutive_failures: int) -> bool:
